Support Rolling Update (#10)
* wip

* add command

* wip

* wip

* export httpHealthCheck

* export LBtimeoutSec

* change

* fix broken test

* fix broken test

* yapf

* yapf

* impl

+++ nata/resources.py
@@ -1,21 +1,24 @@
 from .drivers import *
 import json
 
-class LbResource(object):
 
-    def __init__(self, lb, app):
+class LbResource(object):
+    def __init__(self, lb, app, enable_https=False):
         self.lb = lb
         self.name = lb.name
         self.spec = self.lb.spec
         self.app = app
+        self.enable_https = enable_https
 
     def remote_exist_resources(self):
         result = []
 
-        if exist_global_forwarding_rule(compute, self.name, self.spec, port=80):
+        if exist_global_forwarding_rule(
+                compute, self.name, self.spec, port=80):
             result.append('global_forwarding_rule_http')
 
-        if exist_global_forwarding_rule(compute, self.name, self.spec, port=443):
+        if exist_global_forwarding_rule(
+                compute, self.name, self.spec, port=443):
             result.append('global_forwarding_rule_https')
 
         if exist_global_address(compute, self.name, self.spec):
@@ -39,13 +42,14 @@ class LbResource(object):
         return result
 
     def exist(self):
-        return len(self.remote_exist_resources()) >= 8
+        return len(self.remote_exist_resources()) >= 6
 
     def create(self):
         op = create_http_health_check(compute, self.name, self.spec)
         wait_for_operation(compute, self.spec.project, None, op["name"])
 
-        op = create_backend_service(compute, self.name, self.app.name, self.spec)
+        op = create_backend_service(compute, self.name, self.app.name,
+                                    self.spec)
         wait_for_operation(compute, self.spec.project, None, op["name"])
 
         op = create_url_map(compute, self.name, self.spec)
@@ -54,8 +58,9 @@ class LbResource(object):
         op = create_target_http_proxy(compute, self.name, self.spec)
         wait_for_operation(compute, self.spec.project, None, op["name"])
 
-        op = create_target_https_proxy(compute, self.name, self.spec)
-        wait_for_operation(compute, self.spec.project, None, op["name"])
+        if self.enable_https:
+            op = create_target_https_proxy(compute, self.name, self.spec)
+            wait_for_operation(compute, self.spec.project, None, op["name"])
 
         op = create_global_address(compute, self.name, self.spec)
         wait_for_operation(compute, self.spec.project, None, op["name"])
@@ -63,19 +68,26 @@ class LbResource(object):
         ipaddr = get_global_address(compute, self.name, self.spec)
         self.lb.address = ipaddr
 
-        op = create_global_forwarding_rule(compute, self.name, self.spec, ipaddr, port=80)
+        op = create_global_forwarding_rule(
+            compute, self.name, self.spec, ipaddr, port=80)
         wait_for_operation(compute, self.spec.project, None, op["name"])
 
-        op = create_global_forwarding_rule(compute, self.name, self.spec, ipaddr, port=443)
-        wait_for_operation(compute, self.spec.project, None, op["name"])
-
-    def delete(self):
-        if exist_global_forwarding_rule(compute, self.name, self.spec, port=80):
-            op = delete_global_forwarding_rule(compute, self.name, self.spec, port=80)
+        if self.enable_https:
+            op = create_global_forwarding_rule(
+                compute, self.name, self.spec, ipaddr, port=443)
             wait_for_operation(compute, self.spec.project, None, op["name"])
 
-        if exist_global_forwarding_rule(compute, self.name, self.spec, port=443):
-            op = delete_global_forwarding_rule(compute, self.name, self.spec, port=443)
+    def delete(self):
+        if exist_global_forwarding_rule(
+                compute, self.name, self.spec, port=80):
+            op = delete_global_forwarding_rule(
+                compute, self.name, self.spec, port=80)
+            wait_for_operation(compute, self.spec.project, None, op["name"])
+
+        if self.enable_https and exist_global_forwarding_rule(
+                compute, self.name, self.spec, port=443):
+            op = delete_global_forwarding_rule(
+                compute, self.name, self.spec, port=443)
             wait_for_operation(compute, self.spec.project, None, op["name"])
 
         if exist_global_address(compute, self.name, self.spec):
@@ -86,7 +98,8 @@ class LbResource(object):
             op = delete_target_http_proxy(compute, self.name, self.spec)
             wait_for_operation(compute, self.spec.project, None, op["name"])
 
-        if exist_target_https_proxy(compute, self.name, self.spec):
+        if self.enable_https and exist_target_https_proxy(
+                compute, self.name, self.spec):
             op = delete_target_https_proxy(compute, self.name, self.spec)
             wait_for_operation(compute, self.spec.project, None, op["name"])
 
@@ -113,6 +126,43 @@ class AppResource(object):
         self.name = app.name
         self.spec = self.app.spec
 
+    def get_health_in_backend_service(self, lb):
+        for _ in range(5):
+            health_status = get_health_in_backend_service(compute, self.name,
+                                                      lb.name, self.spec)
+            if 'healthStatus' in health_status:
+                return health_status
+            time.sleep(60)
+
+    def count_stable_instances(self, lb):
+        health_status = self.get_health_in_backend_service(lb)
+        cnt = 0
+        for health in health_status['healthStatus']:
+            if health['healthState'] == 'HEALTHY':
+                cnt += 1
+        return cnt
+
+    def wait_for_stable(self, lb, desire_stable_count):
+        while self.count_stable_instances(lb) != desire_stable_count:
+            print 'wait some maneged instances until stable... desire: %d' % desire_stable_count
+            import time
+            time.sleep(5)
+
+    def rolling(self, lb):
+
+        result = self.get_health_in_backend_service(lb)
+        desire_stable_count = self.count_stable_instances(lb)
+
+        if 'healthStatus' not in result:
+            raise Exception
+
+        for instance in result['healthStatus']:
+            self.wait_for_stable(lb, desire_stable_count)
+            instance_name = instance['instance'].split('/')[-1]
+            op = delete_instance(compute, instance_name, self.spec)
+            wait_for_operation(compute, self.spec.project, self.spec.zone,
+                               op["name"])
+
     def remote_exist_resources(self):
         result = []
 
@@ -132,19 +182,22 @@ class AppResource(object):
 
     def create(self):
         exists_resources = self.remote_exist_resources()
-        if len(exists_resources) != 0 :
-            raise Exception('Already resources exists : {resources}'.format(resources=", ".join(exists_resources)))
+        if len(exists_resources) != 0:
+            raise Exception('Already resources exists : {resources}'.format(
+                resources=", ".join(exists_resources)))
 
         op = create_instance_template(compute, self.name, self.spec)
         wait_for_operation(compute, self.spec.project, None, op["name"])
         print('created instance_template: %s' % self.name)
 
         op = create_instance_group_manager(compute, self.name, self.spec)
-        wait_for_operation(compute, self.spec.project, self.spec.zone, op["name"])
+        wait_for_operation(compute, self.spec.project, self.spec.zone,
+                           op["name"])
         print('created instance_group_maneger: %s' % self.name)
 
         op = create_autoscaler(compute, self.name, self.spec)
-        wait_for_operation(compute, self.spec.project, self.spec.zone, op["name"])
+        wait_for_operation(compute, self.spec.project, self.spec.zone,
+                           op["name"])
         print('created autoscaler: %s' % self.name)
 
     def delete(self):
@@ -153,16 +206,17 @@ class AppResource(object):
 
         if exist_autoscaler(compute, self.name, self.spec):
             op = delete_autoscaler(compute, self.name, self.spec)
-            wait_for_operation(compute, self.spec.project, self.spec.zone, op["name"])
+            wait_for_operation(compute, self.spec.project, self.spec.zone,
+                               op["name"])
             print('deleted autoscaler: %s' % self.name)
 
         if exist_instance_group_manager(compute, self.name, self.spec):
             op = delete_instance_group_manager(compute, self.name, self.spec)
-            wait_for_operation(compute, self.spec.project, self.spec.zone, op["name"])
+            wait_for_operation(compute, self.spec.project, self.spec.zone,
+                               op["name"])
             print('deleted instance_group_maneger: %s', self.name)
 
         if exist_instance_template(compute, self.name, self.spec):
             op = delete_instance_template(compute, self.name, self.spec)
             wait_for_operation(compute, self.spec.project, None, op["name"])
             print('deleted instance_template: %s', self.name)
-
